requirements.txt
Flask==2.3.3
Flask_Cors==4.0.0
Jinja2==3.1.2
numpy==1.24.3
Pillow==10.0.0
Pillow==10.0.1
python_box==7.1.1
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
tensorflow==2.13.0
tensorflow_hub==0.14.0
tensorflow_intel==2.13.0
